Fix mask assertion failing due to moving signed-bit
This is one of the drawbacks of Java's refusal to support unsigned integers: casting a value from one integer type to another actually changes its binary structure.

+++ mapsync-mod/src/main/java/gjum/minecraft/mapsync/mod/data/BlockInfo.java
@@ -7,7 +7,7 @@ import net.minecraft.world.level.block.state.BlockState;
 
 public record BlockInfo(int y, BlockState state) {
 	public void write(BufferWriter writer) throws Exception {
-		writer.writeInt16(y);
+		writer.writeInt16((short) y);
 		writer.writeUnt16(Block.BLOCK_STATE_REGISTRY.getId(state)); // we can assume this never becomes large enough to overflow
 	}
